+++ src/sloths/__init__.py
@@ -4,7 +4,7 @@ Lazy generator pipelines in Python.
 
 from ._stream import Peekable, Stream
 
-__version__ = "0.0.4"
+__version__ = "0.0.5"
 __all__ = (
     "Peekable",
     "Stream",